% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{bc_season_bp_op1}
\alias{bc_season_bp_op1}
\title{Get boxplots of carcasses body condition as a function of season, faceted by season}
\usage{
bc_season_bp_op1(tab, save = FALSE)
}
\arguments{
\item{tab}{table with fat rates translated in body condition}

\item{save}{whether we wish the figure to be saved}
}
\value{
return a boxplot figure
}
\description{
Get boxplots of carcasses body condition as a function of season, faceted by season
}
