% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{bc_season_bp_op1_sp}
\alias{bc_season_bp_op1_sp}
\title{Get species boxplots of carcasses body condition as a function of season, faceted by season}
\usage{
bc_season_bp_op1_sp(tab, sp, save = FALSE)
}
\arguments{
\item{tab}{Table with fat rate converted into a categorical body condition variable}

\item{sp}{the species of interest}

\item{save}{whether we want the graph to be saved or not, default = FALSE}
}
\value{
a bar plot faceted per season for the species of interest
}
\description{
Get species boxplots of carcasses body condition as a function of season, faceted by season
}
