% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{bc_season_bp_op2}
\alias{bc_season_bp_op2}
\title{Get boxplots of carcasses body condition as a function of season, faceted by season}
\usage{
bc_season_bp_op2(tab, save = FALSE)
}
\arguments{
\item{tab}{tab with fat rates translated into body condition}

\item{savewhether}{we wish the figure to be saved}
}
\value{
a boxplot figure
}
\description{
Get boxplots of carcasses body condition as a function of season, faceted by season
}
