% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{bc_season_bp_op2_sp}
\alias{bc_season_bp_op2_sp}
\title{Get species boxplot of carcasses body condition as a function of season, faceted by body condition}
\usage{
bc_season_bp_op2_sp(tab, sp, save = FALSE)
}
\arguments{
\item{tab}{Table with fat rate converted into a categorical body condition variable}

\item{sp}{The species of interest}

\item{save}{Whether you want to save the figure or not}
}
\value{
a barplot faceted per body condition for the species of interest
}
\description{
Get species boxplot of carcasses body condition as a function of season, faceted by body condition
}
