% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{boxplot_main_sp}
\alias{boxplot_main_sp}
\title{Get boxplots of fate rate value or mean fat rate as a function of cluster_month or season, for a subset of species}
\usage{
boxplot_main_sp(
  tab,
  varx,
  vary,
  species_vec = c("buffalo", "nyala", "warthog")
)
}
\arguments{
\item{tab}{the table with fat data}

\item{varx}{explanatory variable (e.g. cluster_month, season1, season 2,...), character string !}

\item{vary}{variable to be explained (?, eg. sample_replicate_fat_rate, mean_fat_rate), character string !}

\item{species_vec}{vector of species of interest, default = c('buffalo', 'nyala', 'warthog')}
}
\value{
return a boxplot figure of the variable to explain ~ explanatory variable facetted by carcass_species
}
\description{
Get boxplots of fate rate value or mean fat rate as a function of cluster_month or season, for a subset of species
}
