% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{chi_test}
\alias{chi_test}
\title{Run chi square test on body condition ~ of season}
\usage{
chi_test(tab, sp = NULL)
}
\arguments{
\item{tab}{table with fat data converted into body condition (categorical variable)}

\item{sp}{default = NULL, must indicate the species to run stat test over a given species}
}
\value{
the contingency table & chi-square results
}
\description{
Run chi square test on body condition ~ of season
}
