% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_data_analyses.R
\name{combine_data}
\alias{combine_data}
\title{Combine data_use and data_available into one dataset for Jacob index calculation and figures}
\usage{
combine_data(data_use, data_av, save = FALSE, species = FALSE)
}
\arguments{
\item{save}{whether we want to save the table, default = F}

\item{species}{TRUE or FALSE, TRUE when combining dataset with species details, otherwise FALSE}

\item{tab_use}{table with the number and proportion made by 'used' prey individual per season and body condition category}

\item{tab_av}{table with the number and proportion made by 'available' prey individual per season and body condition category}
}
\value{
a table with the number and proportion made by 'used' and 'available' prey individuals per season and body condition category
}
\description{
Combine data_use and data_available into one dataset for Jacob index calculation and figures
}
