% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{fat_rate_bone_fresh_bp}
\alias{fat_rate_bone_fresh_bp}
\title{Get boxplots comparing replicate fat rates as function of bone 'freshness', facetted per carcass_ID}
\usage{
fat_rate_bone_fresh_bp(tab, save = FALSE)
}
\arguments{
\item{tab}{table with both sample tested at normal time, and tested according different waiting times}

\item{save}{whether we want to save the figure or not, default = F}
}
\value{
a boxplot comparing replicate fat rates as function of bone 'freshness', facetted per carcass_ID
}
\description{
Get boxplots comparing replicate fat rates as function of bone 'freshness', facetted per carcass_ID
}
