% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{get_body_condition_data_gs}
\alias{get_body_condition_data_gs}
\title{Get 'gold standard' body condition data, i.e. keep adult and old individuals only, with bones collected before 6 days}
\usage{
get_body_condition_data_gs(tab = clean_fat_data, save = FALSE)
}
\arguments{
\item{tab}{table containing (cleaned) fat data, default = clean_fat_data}

\item{save}{whether we wish the table to be saved or not}
}
\value{
return a table with mean fat rates translated into body condition, 3 levels : emaciated (>0.20), thin (between 0.20 & 0.70) & good (>0.70)
}
\description{
Get 'gold standard' body condition data, i.e. keep adult and old individuals only, with bones collected before 6 days
}
