% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{get_CV_fat_rate_table}
\alias{get_CV_fat_rate_table}
\title{Get fat rate coefficient of variation for each samples (i.e. per sample_ID)}
\usage{
get_CV_fat_rate_table(tab = clean_fat_data, save = T)
}
\arguments{
\item{tab}{fat data cleaned with 'clean_raw_fat_data' function, default is clean_fat_data}

\item{save}{whether we want the table to be saved or not}
}
\value{
return a table with the nb of replicates, mean fat rate, SD fat rate and Coefficient of variation (CV) for each sample(i.e. Sample_ID)
}
\description{
Get fat rate coefficient of variation for each samples (i.e. per sample_ID)
}
