% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_data_analyses.R
\name{get_data_available}
\alias{get_data_available}
\title{Get table of body condition of prey available in the landscape}
\usage{
get_data_available(data_av, save = FALSE, species = FALSE)
}
\arguments{
\item{data_av}{a table with fat data and mean fat rate translated into categorical body conditions of available prey}

\item{save}{whether we want to save the table or not (if so, will be saved in the 'analyses' folder), default = F}

\item{species}{TRUE or FALSE, if TRUE : compute number and proportion made by 'used' prey individual per season and body condition category and species}
}
\value{
a table with the number and proportion made by 'available' prey individual per season and body condition category
}
\description{
Get table of body condition of prey available in the landscape
}
