% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_data_analyses.R
\name{get_data_use}
\alias{get_data_use}
\title{Get table of prey used (i.e. killed) by lions}
\usage{
get_data_use(data_use, save = FALSE, species = T)
}
\arguments{
\item{data_use}{a table with fat data and mean fat rate translated into categorical body conditions of 'used' prey}

\item{save}{whether we want to save the table or not (if so, will be saved in the 'analyses' folder), default = F}

\item{species}{TRUE or FALSE, if TRUE : compute number and proportion made by 'used' prey individual per season and body condition category and species}
}
\value{
a table with the number and proportion made by 'used' prey individual per season and body condition category
}
\description{
Get table of prey used (i.e. killed) by lions
}
