% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{get_dry_mass_boxplot}
\alias{get_dry_mass_boxplot}
\title{Get boxplots of dry masses for each samples (i.e. sample_ID)}
\usage{
get_dry_mass_boxplot(tab = clean_fat_data, save = T)
}
\arguments{
\item{save}{whether we want the figure to be saved or not}

\item{clean_fat_data}{fat data cleaned with 'clean_raw_fat_data' function}
}
\value{
boxtplots of dry masses ~ sample_ID
}
\description{
Get boxplots of dry masses for each samples (i.e. sample_ID)
}
