% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{get_fat_rate_boxplot}
\alias{get_fat_rate_boxplot}
\title{Get boxplots of fat rates for each samples (i.e. sample_ID)}
\usage{
get_fat_rate_boxplot(tab = clean_fat_data, save = TRUE)
}
\arguments{
\item{tab}{fat data cleaned with 'clean_raw_fat_data' function, default is clean_fat_data}

\item{save}{whether we want the figure to be saved or not}
}
\value{
boxtplots of fat rates as a function of/for each sample_ID
}
\description{
Get boxplots of fat rates for each samples (i.e. sample_ID)
}
