% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{get_fat_rate_CV_hist}
\alias{get_fat_rate_CV_hist}
\title{Get an histogram of fat rate coefficients of variation}
\usage{
get_fat_rate_CV_hist(tab, save = F)
}
\arguments{
\item{tab}{table with fat rate coefficients of variation}

\item{save}{whether we want the figure to be saved or not, default = F}
}
\value{
a histogram of fat rate coeeficient of variation
}
\description{
Get an histogram of fat rate coefficients of variation
}
