% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{get_fat_rate_evol_data}
\alias{get_fat_rate_evol_data}
\title{Get data set to explore fat rate change with increasing analysis waiting time}
\usage{
get_fat_rate_evol_data(tab, save = FALSE)
}
\arguments{
\item{tab}{table with raw fat data (i.e. not cleaned yet), should be fat_data_raw}

\item{save}{whether we want the table to be save or not, default = F}
}
\value{
a table with both sample tested at normal time, and tested according different waiting times
}
\description{
Get data set to explore fat rate change with increasing analysis waiting time
}
