% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_figures.R
\name{get_fig_barplot}
\alias{get_fig_barplot}
\title{Get barplot figure of proportion of used and available prey in each body condition category per season, for all species or for a given species}
\usage{
get_fig_barplot(data_tab, save = T, species = NULL)
}
\arguments{
\item{data_tab}{table with number of individual in each body condition catagory}

\item{save}{whether we want to save the figure or not (in 'figure' directory)}

\item{species}{if NULL barplot is draw for all species, if given a species, barplot is draw for this species}
}
\value{
return a barplot of proportion of used and available prey in each body condition category per season
}
\description{
Get barplot figure of proportion of used and available prey in each body condition category per season, for all species or for a given species
}
