% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_figures.R
\name{get_fig_ratio}
\alias{get_fig_ratio}
\title{Get the graphical representation of Jacob's indices per season}
\usage{
get_fig_ratio(ratio_table, save = F, species = NULL)
}
\arguments{
\item{ratio_table}{table with Jacob's indices computed across all species or for a given species}

\item{save}{whether we want to save the figure}

\item{species}{if we compute the figure for a given species, must be indicated}
}
\value{
a figure corresponding to a graphical representation of Jacob's indices
}
\description{
Get the graphical representation of Jacob's indices per season
}
