% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_figures.R
\name{get_figA2_fat_rate_bone_fresh}
\alias{get_figA2_fat_rate_bone_fresh}
\title{Get boxplots comparing replicate fat rates as function of bone 'freshness', facetted per carcass_ID}
\usage{
get_figA2_fat_rate_bone_fresh(tab, save = FALSE)
}
\arguments{
\item{tab}{table with both sample tested at normal time, and tested according different waiting times}

\item{save}{whether we want to save the figure or not, default = F}
}
\value{
a boxplot comparing replicate fat rates as function of bone 'freshness', facetted per carcass_ID
}
\description{
Get boxplots comparing replicate fat rates as function of bone 'freshness', facetted per carcass_ID
}
