% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{get_gs_data}
\alias{get_gs_data}
\title{Get 'gold standard' data, i.e. keep adult and old individuals only, with bones collected before 6 days}
\usage{
get_gs_data(tab = clean_fat_data, save = FALSE)
}
\arguments{
\item{tab}{table containing (cleaned) fat data, default = clean_fat_data}

\item{save}{whether we want the table to be saved, default = FALSE}
}
\value{
return a table
}
\description{
Get 'gold standard' data, i.e. keep adult and old individuals only, with bones collected before 6 days
}
