% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{get_high_CV_table}
\alias{get_high_CV_table}
\title{Get a table with the higher coefficient of variation}
\usage{
get_high_CV_table(tab, value = 0.25, CV_fat_rate_table, save = FALSE)
}
\arguments{
\item{tab}{table with clean fat data}

\item{value}{the value above which a CV is considered high, default = 0.25}

\item{save}{whether we want the table to be saved of not, default = F}
}
\value{
a table of the samples with fat rate coefficient of variation (strictly) higher than 'value'
}
\description{
Get a table with the higher coefficient of variation
}
