% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_figures.R
\name{get_high_CV_table_simplified}
\alias{get_high_CV_table_simplified}
\title{Get a simplified table of samples with high coefficient of variation}
\usage{
get_high_CV_table_simplified(tab = high_CV_table, save = F)
}
\arguments{
\item{tab}{table with the samples with high coefficient of variation}

\item{save}{wether we want to save the table or not, saved in figure/appendices/}
}
\value{
a table with as many row as the number of samples with high CV
}
\description{
Get a simplified table of samples with high coefficient of variation
}
