% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{get_ls_data}
\alias{get_ls_data}
\title{Get larger sample data, i.e. keep sub-adult, adult & old individuals only, with bones collected before 6 days or between 6  & 8 days}
\usage{
get_ls_data(tab = clean_fat_data, save = FALSE)
}
\arguments{
\item{tab}{table with cleaned fat data}

\item{save}{whether we want the data to be saved or not}
}
\value{
a table with larger sample data
}
\description{
Get larger sample data, i.e. keep sub-adult, adult & old individuals only, with bones collected before 6 days or between 6  & 8 days
}
