% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_data_analyses.R
\name{get_ratio_table}
\alias{get_ratio_table}
\title{Get table with jacob indices}
\usage{
get_ratio_table(data_tab, save = F, species = NULL)
}
\arguments{
\item{data_tab}{dataset combining used and available prey body condition data and proportion in each catagory}

\item{save}{whether we want to save the table, default = F}

\item{species}{must specified the species if ones want ratio table for a given species, otherwise NULL (default) give ratio across species}
}
\value{
a table with proportion of used and available prey, and jacob index J_I(used,av) --> selection/avoidance of used in comparison with the available
}
\description{
Get table with jacob indices
}
