% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_figures.R
\name{get_summary_table}
\alias{get_summary_table}
\title{Get data summary table of the number of carcasses for each species, per season}
\usage{
get_summary_table(data_use, save = T)
}
\arguments{
\item{data_use}{table with the detail of species, season}

\item{save}{whether we want to save the table or not, default = T, saved in 'figure' file}
}
\value{
return a table with total number of carcasses and the number of carcasses per season, and the totals
}
\description{
Get data summary table of the number of carcasses for each species, per season
}
