% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_data_analyses.R
\name{J_I}
\alias{J_I}
\title{Compute Jacob's index}
\usage{
J_I(a, b)
}
\arguments{
\item{a}{proportion of prey used (i.e. killed) by lions}

\item{b}{proportion of prey available (for comparison)}
}
\value{
D, jacob index value indicating of a selection/avoidance of a over b (?), standardized, range from -1 (max avoidance) to +1 (maximum preference)
}
\description{
Compute Jacob's index
}
