% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{mean_fat_rate_bone_fresh_bp}
\alias{mean_fat_rate_bone_fresh_bp}
\title{Get scatterplot showing mean fat rates as function of bone 'freshness', facetted per carcass_ID}
\usage{
mean_fat_rate_bone_fresh_bp(tab, save = F)
}
\arguments{
\item{tab}{table with both sample tested at normal time, and tested according different waiting times}

\item{save}{whether we want the figure to be saved or not, default = F}
}
\value{
a scatterplot of mean fat rate as a function of bone 'freshness', facetted by carcass_ID
}
\description{
Get scatterplot showing mean fat rates as function of bone 'freshness', facetted per carcass_ID
}
