% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_read_dataset.R
\name{read_data_land_BC_sp}
\alias{read_data_land_BC_sp}
\title{Function aims at reading available prey body condition data for comparison with body condition used (i.e. preyed) by lions, per species}
\usage{
read_data_land_BC_sp()
}
\value{
a table with the number of individual in each BC catagories (column) per seasons (raws)
}
\description{
Function aims at reading available prey body condition data for comparison with body condition used (i.e. preyed) by lions, per species
}
