% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_read_dataset.R
\name{read_fat_data_raw}
\alias{read_fat_data_raw}
\title{Read raw data for fat analyses, from Yolan (after inclusion of collection_interval column)}
\usage{
read_fat_data_raw()
}
\value{
raw fat dataset
}
\description{
Read raw data for fat analyses, from Yolan (after inclusion of collection_interval column)
}
