% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_read_dataset.R
\name{read_fat_data_raw_2}
\alias{read_fat_data_raw_2}
\title{Read updated raw data for fat analyses, from Yolan, sent the 8th June 2023 (after inclusion of collection_interval column and removal of unnacessary raws (see read_me, data preparation))}
\usage{
read_fat_data_raw_2()
}
\value{
updated 8th june 2023 raw fat dataset
}
\description{
Read updated raw data for fat analyses, from Yolan, sent the 8th June 2023 (after inclusion of collection_interval column and removal of unnacessary raws (see read_me, data preparation))
}
