% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_clean_dataset.R
\name{save_cleaned_data}
\alias{save_cleaned_data}
\title{Save cleaned data for fat analyses}
\usage{
save_cleaned_data(cleaned_data, filename)
}
\arguments{
\item{cleaned_data}{cleaned raw data for fat analyses (after using clean_raw_fat_data function)}

\item{filename}{the name we want the dataset to be saved under}
}
\value{
save cleaned fat data
}
\description{
Save cleaned data for fat analyses
}
