% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_exploration.R
\name{wilcox_mann_whitney}
\alias{wilcox_mann_whitney}
\title{Run Wilcoxon-Mann-Whitney statistical test on mean fat rate for either all species or specific species}
\usage{
wilcox_mann_whitney(tab, sp)
}
\arguments{
\item{tab}{table with replicate fat rate data, either gs_data (gold standard data) or ls-data (larger sample data)}

\item{sp}{: NULL for all species, otherwise the species must be specified}
}
\value{
3 figures (plot, boxplot and scatterplots), a data summary, results of shapiro test, results of Wilcoxon Mann Whitney
}
\description{
Run Wilcoxon-Mann-Whitney statistical test on mean fat rate for either all species or specific species
}
