% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_clean_dataset.R
\name{clean_raw_fat_data}
\alias{clean_raw_fat_data}
\title{Clean raw data for fat analyses}
\usage{
clean_raw_fat_data(fat_data_raw)
}
\arguments{
\item{fat_data_raw}{raw data for fat analyses obtain from the read with read_data function}
}
\value{
cleaned fat data
}
\description{
Clean raw data for fat analyses
}
